% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scores.R
\name{AdjScoreClustersCODEX}
\alias{AdjScoreClustersCODEX}
\title{Use the Adjacency Score to evaluate colocaliztion of all pairs
of CODEX clusters in CODEX spatial dimensions.
Calls Adjacency Score with c=0 to use hypergeometric null distribution
speed up for mutually exclusive binary features.}
\usage{
AdjScoreClustersCODEX(stvea_object, k, num_cores = 1)
}
\arguments{
\item{stvea_object}{STvEA.data class object containing CODEX
cluster assignments and CODEX spatial coordinates}

\item{k}{number of nearest neighbors to create graph
from CODEX spatial information}

\item{num_cores}{integer specifying the number of cores to be used
in the computation. By default only one core is used.}
}
\description{
Use the Adjacency Score to evaluate colocaliztion of all pairs
of CODEX clusters in CODEX spatial dimensions.
Calls Adjacency Score with c=0 to use hypergeometric null distribution
speed up for mutually exclusive binary features.
}
