% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scores.R
\name{AdjScoreGenes.internal}
\alias{AdjScoreGenes.internal}
\title{Use the Adjacency Score to evaluate colocalization of given
pairs of genes mapped to the CODEX spatial positions
Takes matrices and data frames instead of STvEA.data class}
\usage{
AdjScoreGenes.internal(adj_matrix, codex_mRNA, gene_pairs, c = 0,
  num_cores = 1, num_perms = 1000, perm_estimate = T)
}
\arguments{
\item{adj_matrix}{a (preferrably sparse) binary matrix of
adjacency between the cells in the CODEX spatial coordinates}

\item{codex_mRNA}{a (CODEX cells x genes) matrix of the CITE-seq
mRNA expression mapped to the CODEX cells}

\item{gene_pairs}{a 2 column matrix of gene pairs where each row
specifies the names of the genes in a pair}

\item{c}{constant used to determine width of diffusion, must be 0 <= c}

\item{num_cores}{integer specifying the number of cores to be used
in the computation. By default only one core is used.}

\item{num_perms}{number of permutations used to build the null
distribution for each feature. By default is set to 1000.}

\item{perm_estimate}{boolean indicating whether Gaussian distribution
parameters should be determined from num_perms permutations to estimate
the p-value. By default is set to TRUE.}
}
\description{
Use the Adjacency Score to evaluate colocalization of given
pairs of genes mapped to the CODEX spatial positions
Takes matrices and data frames instead of STvEA.data class
}
