% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{AnchorCorrection}
\alias{AnchorCorrection}
\title{Run all methods for anchor correction}
\usage{
AnchorCorrection(
  ref_mat,
  query_mat,
  rna_mat,
  cite_index,
  num.cc = NULL,
  k.anchor = 20,
  k.filter = 100,
  k.score = 80,
  k.weight = 100,
  verbose = FALSE
)
}
\arguments{
\item{ref_mat}{a (cell x feature) protein expression matrix}

\item{query_mat}{a (cell x feature) protein expression matrix to be corrected}

\item{rna_mat}{a (cell x feature) embedding of the mRNA expression matrix from CITE-seq}

\item{cite_index}{which matrix (1 or 2) is the CITE-seq protein expression matrix}

\item{num.cc}{number of canonical vectors to calculate. Defaults to number of proteins - 1}

\item{k.anchor}{number of nn used to find anchors via mutual nearest neighbors}

\item{k.filter}{number of nn in original feature space to use for filtering}

\item{k.score}{number of nn to use in shared nearest neighbor scoring}

\item{k.weight}{number of nn in original query feature space to make correction vectors}

\item{verbose}{print status at each step}
}
\description{
Run all methods for anchor correction
}
