% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{clean_cite}
\alias{clean_cite}
\title{Removes noise from CITE-seq protein data by fitting
a Negative Binomial mixture and computing each expression
measurement as the cumulative distribution of the
Negative Binomial with the higher median.}
\usage{
clean_cite(cite_protein, maxit = 500, factr = 1e-09,
  optim_inits = NULL, num_cores = 1)
}
\arguments{
\item{cite_protein}{Raw CITE-seq protein data (cell x protein)}

\item{maxit}{maximum number of iterations for optim function}

\item{factr}{accuracty of optim function}

\item{optim_inits}{a matrix of (proteins x params) with initialization
parameters for each protein to input to the optim function. If NULL,
starts at two default parameter sets and picks the better one}

\item{verbose}{print out each iteration of negative binomial fit loop}
}
\value{
Cleaned CITE-seq protein data matrix (cell x protein)
}
\description{
Removes noise from CITE-seq protein data by fitting
a Negative Binomial mixture and computing each expression
measurement as the cumulative distribution of the
Negative Binomial with the higher median.
}
