% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{clean_codex}
\alias{clean_codex}
\title{Removes noise from CODEX data by fitting a Gaussian
mixture and computing each expression measurement
as the cumulative distribution of the Gaussian with
the higher mean.}
\usage{
clean_codex(codex_filtered)
}
\arguments{
\item{codex_filtered}{CODEX expression matrix after filtering (cell x proteins)}
}
\value{
Cleaned CODEX protein expression matrix (cell x protein)
}
\description{
Removes noise from CODEX data by fitting a Gaussian
mixture and computing each expression measurement
as the cumulative distribution of the Gaussian with
the higher mean.
}
