% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{CleanCODEX}
\alias{CleanCODEX}
\title{Removes noise from CODEX data by fitting a Gaussian
mixture and computing each expression measurement
as the cumulative distribution of the Gaussian with
the higher mean.}
\usage{
CleanCODEX(stvea_object)
}
\arguments{
\item{stvea_object}{STvEA.data class with CODEX data after FilterCODEX}
}
\description{
Removes noise from CODEX data by fitting a Gaussian
mixture and computing each expression measurement
as the cumulative distribution of the Gaussian with
the higher mean.
}
