% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ClusterCITE.internal}
\alias{ClusterCITE.internal}
\title{Perform HDBSCAN consensus clustering on CITE-seq latent space}
\usage{
ClusterCITE.internal(cite_latent)
}
\description{
Perform HDBSCAN consensus clustering on CITE-seq latent space
}
