% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ClusterCITE}
\alias{ClusterCITE}
\title{Perform HDBSCAN consensus clustering on CITE-seq latent space}
\usage{
ClusterCITE(cite_latent)
}
\description{
Perform HDBSCAN consensus clustering on CITE-seq latent space
}
