% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ClusterCODEX.internal}
\alias{ClusterCODEX.internal}
\title{Perform louvain clustering on CODEX data
Takes matrices and data frames instead of STvEA.data class}
\usage{
ClusterCODEX.internal(codex_knn, k = ncol(codex_knn))
}
\arguments{
\item{codex_knn}{matrix of (codex cells x k) nearest neighbor indices
in the CODEX protein space}

\item{k}{number of nearest neighbors to use in graph-based clustering}
}
\description{
Perform louvain clustering on CODEX data
Takes matrices and data frames instead of STvEA.data class
}
