% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{clustering_names}
\alias{clustering_names}
\title{Assign user input labels to CITE-seq clusters for future plotting}
\usage{
clustering_names(labels)
}
\description{
Assign user input labels to CITE-seq clusters for future plotting
}
