% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ConsensusCluster.internal}
\alias{ConsensusCluster.internal}
\title{Keep only HDBSCAN results that pass a certain silhouette score cutoff
and create a dissimilarity matrix between cells from the clusterings
Perform agglomerative hierarchical clustering on the
consensus dissimilarity matrix
Takes matrices and data frames instead of STvEA.data class}
\usage{
ConsensusCluster.internal(hdbscan_results, cite_latent, silhouette_cutoff,
  num_cluster)
}
\description{
Keep only HDBSCAN results that pass a certain silhouette score cutoff
and create a dissimilarity matrix between cells from the clusterings
Perform agglomerative hierarchical clustering on the
consensus dissimilarity matrix
Takes matrices and data frames instead of STvEA.data class
}
