% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ConsensusCluster}
\alias{ConsensusCluster}
\title{Perform agglomerative hierarchical clustering on the
dissimilarity matrix output by GetConsensusMatrix}
\usage{
ConsensusCluster(dissim_matrix, num_cluster)
}
\description{
Perform agglomerative hierarchical clustering on the
dissimilarity matrix output by GetConsensusMatrix
}
