% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_anchor_correction.R
\name{ConstructNNMat}
\alias{ConstructNNMat}
\title{Construct nearest neighbor matrix from nn.idx}
\usage{
ConstructNNMat(nn.idx, offset1, offset2, dims)
}
\arguments{
\item{nn.idx}{Nearest neighbor index matrix (nn.idx from RANN)}

\item{offset1}{Offsets for the first neighbor}

\item{offset2}{Offsets for the second neighbor}
}
\value{
returns a sparse matrix representing the NN matrix
}
\description{
Construct nearest neighbor matrix from nn.idx
}
