% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{CorNN}
\alias{CorNN}
\title{Find the knn indices in data for all cells in query
using Pearson's correlation dissimilarity}
\usage{
CorNN(data, query = data, k = 5)
}
\arguments{
\item{data}{a (M cell x d feature) matrix}

\item{query}{a (N cell x d feature) matrix. If not specified, query = data.}

\item{k}{number of nearest neighbors to find}
}
\value{
a N x k matrix containing nearest neighbors for each cell in query
}
\description{
Note: "query" here is used to mimic RANN documentation, not Seurat documentation
}
