% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{correctChunksCODEX}
\alias{correctChunksCODEX}
\title{Run all methods for anchor correction}
\usage{
correctChunksCODEX(cite_protein, codex_protein, cite_latent, num_chunks,
  seed = NULL, num.cc = 29, k.anchor = 20, k.filter = 100,
  k.score = 80, k.weight = 100, verbose = FALSE)
}
\arguments{
\item{num_chunks}{number of equal sized chunks to split CODEX dataset into for correction}

\item{num.cc}{number of canonical vectors to calculate}

\item{k.anchor}{number of nn used to find anchors via mutual nearest neighbors}

\item{k.filter}{number of nn in original feature space to use for filtering}

\item{k.score}{number of nn to use in shared nearest neighbor scoring}

\item{k.weight}{number of nn in original query feature space to make correction vectors}

\item{ref_mat}{a (n cell x f feature) protein expression matrix}

\item{query_mat}{a (m cell x f feature) protein expression matrix to be corrected}

\item{rna_mat}{a (cell x feature) embedding of the mRNA expression matrix from CITE-seq}

\item{cite_index}{which matrix (1 or 2) is the CITE-seq protein expression matrix}
}
\value{
a (m cell x f feature) expression matrix of the CODEX data corrected into the CITE-seq space
}
\description{
Run all methods for anchor correction
}
