% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_anchor_correction.R
\name{FilterAnchors}
\alias{FilterAnchors}
\title{Filter anchors, keeping those that are also in the knn
 of the L2Norm of the original feature space}
\usage{
FilterAnchors(ref_mat, query_mat, anchors, k.filter = 200, eps = 0,
  verbose = FALSE)
}
\arguments{
\item{ref_mat}{a (cell x feature) protein expression matrix}

\item{query_mat}{a (cell x feature) protein expression matrix to be corrected}

\item{anchors}{the list of anchors between matrix1 and matrix2 as output by FindAnchorPairs}

\item{k.filter}{number of nearest neighbors in original feature space to use for filtering}

\item{eps}{error bound on the neighbor finding algorithm (from \code{\link{RANN}})}

\item{verbose}{...}
}
\value{
a matrix where each row is an anchor pair of cell indices
}
\description{
Filter anchors, keeping those that are also in the knn
 of the L2Norm of the original feature space
}
