% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{FilterCITEmRNA.internal}
\alias{FilterCITEmRNA.internal}
\title{Filter CITE-seq mRNA data based on number of transcripts per cell
and the number of cells each gene is expressed in
Takes matrices and data frames instead of STvEA.data class}
\usage{
FilterCITEmRNA.internal(
  cite_mRNA,
  min_transcripts = 1200,
  min_cells_per_gene = 30
)
}
\arguments{
\item{cite_mRNA}{a (cell x gene) matrix of CITE-seq mRNA data}

\item{min_transcripts}{minimum number of transcripts to keep cell}

\item{min_cells_per_gene}{minimum number of cells a gene must be
expressed in to be kept}
}
\description{
Filter CITE-seq mRNA data based on number of transcripts per cell
and the number of cells each gene is expressed in
Takes matrices and data frames instead of STvEA.data class
}
