% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{FilterCITEmRNA}
\alias{FilterCITEmRNA}
\title{Filter CITE-seq mRNA data based on number of transcripts per cell
and the number of cells each gene is expressed in}
\usage{
FilterCITEmRNA(stvea_object, min_transcripts = 1200,
  min_cells_per_gene = 30)
}
\arguments{
\item{stvea_object}{STvEA.data class with CITE-seq mRNA data}

\item{min_transcripts}{minimum number of transcripts to keep cell}

\item{min_cells_per_gene}{minimum number of cells a gene must be}
}
\description{
Filter CITE-seq mRNA data based on number of transcripts per cell
and the number of cells each gene is expressed in
}
