% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{FilterCODEX}
\alias{FilterCODEX}
\title{Removes points from CODEX matrix that are not cells
as determined by the gating strategy on the blank channels
from the CODEX paper}
\usage{
FilterCODEX(
  stvea_object,
  size_lim = NULL,
  blank_upper = NULL,
  blank_lower = NULL
)
}
\arguments{
\item{stvea_object}{STvEA.data class with CODEX data}

\item{size_lim}{lower and upper limits on size of each cell. If blank, set to 0.025 and 0.99 quantiles}

\item{blank_upper}{a vector with an upper bound expression cutoff for each blank channel.
If NULL, blank upper bounds are set as the 0.995 quantile for each blank}

\item{blank_lower}{a vector with a lower bound expression cutoff for each blank channel.
If NULL, blank lower bounds are set as the 0.002 quantile for each blank}
}
\description{
Removes points from CODEX matrix that are not cells
as determined by the gating strategy on the blank channels
from the CODEX paper
}
