% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_anchor_correction.R
\name{FindAnchorPairs}
\alias{FindAnchorPairs}
\title{Find Anchor pairs as MNN between two datasets}
\usage{
FindAnchorPairs(neighbors, k.anchor = 5, verbose = FALSE)
}
\arguments{
\item{neighbors}{a list of knn neighbors between two datasets as output by FindNN}

\item{k.anchor}{number of nearest neighbors to look through for anchors}

\item{verbose}{...}
}
\value{
a matrix where each row is an anchor pair of cell indices
}
\description{
Find Anchor pairs as MNN between two datasets
}
