% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_anchor_correction.R
\name{FindIntegrationMatrix}
\alias{FindIntegrationMatrix}
\title{Calculate anchor vectors between reference and query dataset
This is matrix B from Methods in Stuart et al.}
\usage{
FindIntegrationMatrix(ref_mat, query_mat, neighbors, anchors, verbose = FALSE)
}
\arguments{
\item{ref_mat}{a (cell x feature) protein expression matrix}

\item{query_mat}{a (cell x feature) protein expression matrix to be corrected}

\item{neighbors}{a list of neighbors}

\item{anchors}{a list of anchors (MNN) as from FindAnchorPairs}

\item{verbose}{...}
}
\description{
Calculate anchor vectors between reference and query dataset
This is matrix B from Methods in Stuart et al.
}
