% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_anchor_correction.R
\name{FindNNrna}
\alias{FindNNrna}
\title{Find nearest neighbors}
\usage{
FindNNrna(
  ref_emb,
  query_emb,
  rna_mat,
  cite_index = 1,
  k = 300,
  eps = 0,
  verbose = FALSE
)
}
\arguments{
\item{ref_emb}{a (cell x feature) embedding of a protein expression matrix}

\item{query_emb}{a (cell x feature) embedding of protein expression matrix to be corrected}

\item{rna_mat}{a (cell x feature) embedding of the mRNA expression matrix from CITE-seq}

\item{cite_index}{which matrix (1 or 2) is the CITE-seq protein expression matrix}

\item{k}{number of nearest neighbor to find between the matrices}

\item{eps}{error bound on nearest neighbor search (see eps parameter for nn2)}

\item{verbose}{...}
}
\value{
list of nearest neighbor lists:
nnab = for each cell in a, its knn in b
nnba = for each cell in b, its knn in a
nnbb = knn within b, nnaa = knn within a
}
\description{
Find nearest neighbors
}
