% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_anchor_correction.R
\name{FindWeights}
\alias{FindWeights}
\title{Calculates weights of the k nearest anchors for each cell in the query dataset}
\usage{
FindWeights(neighbors, anchors, query_mat, integration.matrix,
  k.weight = 300, sd.weight = 1, eps = 0, verbose = TRUE)
}
\arguments{
\item{neighbors}{list of neighbors in CCA space as output by FindNN}

\item{anchors}{list of anchors as output by FindAnchorPairs or FilterAnchors}

\item{query_mat}{a (cells x features) matrix for the query dataset (protein or mRNA)}

\item{integration.matrix}{the output of FindIntegrationMatrix}

\item{sd.weight}{standard deviation of the Gaussian kernel}

\item{eps}{Error bound on the neighbor finding algorithm (from \code{\link{RANN}})}

\item{verbose}{...}

\item{k}{number of nearest anchors to use in correction}
}
\description{
Calculates weights of the k nearest anchors for each cell in the query dataset
}
