% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{FitGaussian}
\alias{FitGaussian}
\title{Fits the log-normalized expression values of one protein with a Gaussian mixture
Takes matrices and data frames instead of STvEA.data class}
\usage{
FitGaussian(norm_protein_expr)
}
\arguments{
\item{norm_protein_expr}{Log-normalized CITE-seq protein data for one protein}
}
\description{
Fits the log-normalized expression values of one protein with a Gaussian mixture
Takes matrices and data frames instead of STvEA.data class
}
