% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{FitNB}
\alias{FitNB}
\title{Fits the expression values of one protein with a
Negative Binomial mixture}
\usage{
FitNB(protein_expr, maxit = 500, factr = 1e-09, optim_init = NULL)
}
\arguments{
\item{protein_expr}{Raw CITE-seq protein data for one protein}

\item{maxit}{maximum number of iterations for optim function}

\item{factr}{accuracty of optim function}

\item{optim_init}{optional initialization parameters for the optim function
if NULL, starts at two default parameter sets and picks the better one}
}
\description{
Fits the expression values of one protein with a
Negative Binomial mixture
}
