% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_nn_matrix}
\alias{get_nn_matrix}
\title{Get knn in to_dataset for each cell in from_dataset
based on the corrected data matrix from AnchorCorrection}
\usage{
get_nn_matrix(from_dataset, to_dataset, k = floor(nrow(to_dataset) *
  0.002), c = 0.1)
}
\arguments{
\item{from_dataset}{a (cell x feature) matrix or dataframe}

\item{to_dataset}{a (cell x feature) matrix or dataframe}

\item{k}{number of nearest neighbors to find}

\item{c}{constant controls the width of the Gaussian kernel}
}
\description{
Get knn in to_dataset for each cell in from_dataset
based on the corrected data matrix from AnchorCorrection
}
