% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{GetConsensusMatrix}
\alias{GetConsensusMatrix}
\title{Keep only HDBSCAN results that pass a certain silhouette score cutoff
and create a dissimilarity matrix between cells from the clusterings}
\usage{
GetConsensusMatrix(hdbscan_results, silhouette_cutoff)
}
\description{
Keep only HDBSCAN results that pass a certain silhouette score cutoff
and create a dissimilarity matrix between cells from the clusterings
}
