% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{GetMatrixNN}
\alias{GetMatrixNN}
\title{Gets NN matrix for both CITE->CODEX and CODEX->CITE, stores in MappingDataHolder}
\usage{
GetMatrixNN(mapping_object,
  k.codex = floor(nrow(mapping_object@cite_clean) * 0.002),
  k.cite = floor(nrow(mapping_object@corrected_codex) * 0.002),
  c.codex = 0.1, c.cite = 0.1)
}
\arguments{
\item{mapping_object}{MappingDataHolder class with mapping correction run}

\item{k.codex}{number of nearest CITE-seq neighbors to find for each CODEX cell}

\item{k.cite}{number of nearest CODEX neighbors to find for each CITE-seq cell}

\item{c.codex}{width of Gaussian kernel in CODEX dataset}

\item{c.cite}{width of Gaussian kernel in CITE-seq dataset}
}
\description{
Gets NN matrix for both CITE->CODEX and CODEX->CITE, stores in MappingDataHolder
}
