% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{GetTransferMatrix}
\alias{GetTransferMatrix}
\title{Gets NN matrix for CITE->CODEX, stores in STvEA.data}
\usage{
GetTransferMatrix(
  stvea_object,
  k.cite = floor(nrow(stvea_object@corrected_codex) * 0.002),
  c.cite = 0.1,
  transfer.rna = TRUE,
  fastknn = FALSE
)
}
\arguments{
\item{stvea_object}{STvEA.data class with mapping correction run}

\item{k.cite}{number of nearest CODEX neighbors to find for each CITE-seq cell}

\item{c.cite}{width of Gaussian kernel in CITE-seq dataset}

\item{fastknn}{if set to TRUE, skips the computation of the NN matrix and uses the FastKNN library to transfer
the CITE-seq labels into the CODEX data}

\item{transfer_rna}{if set to TRUE (default), also transfers the mRNA data from CITE-seq to CODEX. Requires extra memory}
}
\description{
Gets NN matrix for CITE->CODEX, stores in STvEA.data
}
