% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GetUmapCITE}
\alias{GetUmapCITE}
\title{Runs UMAP on CITE-seq latent space (if provided) or
mRNA expression}
\usage{
GetUmapCITE(stvea_object, ...)
}
\arguments{
\item{stvea_object}{STvEA.data class object with CITE-seq
latent space or mRNA expression}

\item{...}{parameters to pass into umap}
}
\description{
Runs UMAP on CITE-seq latent space (if provided) or
mRNA expression
}
