% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GetUmapCODEX}
\alias{GetUmapCODEX}
\title{Runs UMAP on CODEX protein expression. If CleanCODEX has been run,
calls UMAP on the cleaned CODEX protein expression. Otherwise,
runs UMAP on the original or filtered CODEX protein expression.}
\usage{
GetUmapCODEX(stvea_object, ...)
}
\arguments{
\item{stvea_object}{STvEA.data class object}

\item{...}{parameters to pass into umap}
}
\description{
Runs UMAP on CODEX protein expression. If CleanCODEX has been run,
calls UMAP on the cleaned CODEX protein expression. Otherwise,
runs UMAP on the original or filtered CODEX protein expression.
}
