% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umap.R
\name{GetVisUMAP}
\alias{GetVisUMAP}
\title{Runs UMAP on CITE-seq latent space and CODEX protein expression
if they exist in the provided STvEA.data object. If CleanCODEX
has been run, calls UMAP on the cleaned CODEX protein expression.
Otherwise, runs UMAP on the original CODEX protein expression.}
\usage{
GetVisUMAP(stvea_object, ...)
}
\arguments{
\item{stvea_object}{STvEA.data class object}

\item{...}{parameters to pass into umap}
}
\description{
Runs UMAP on CITE-seq latent space and CODEX protein expression
if they exist in the provided STvEA.data object. If CleanCODEX
has been run, calls UMAP on the cleaned CODEX protein expression.
Otherwise, runs UMAP on the original CODEX protein expression.
}
