% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{KnnCODEX}
\alias{KnnCODEX}
\title{Get k nearest neighbors for CODEX protein space. If CleanCODEX
has been run, uses cleaned CODEX protein data, otherwise uses
original CODEX protein data.}
\usage{
KnnCODEX(stvea_object, k = 5)
}
\arguments{
\item{stvea_object}{STvEA.data class object with CODEX protein data}

\item{k}{number of nearest neighbors to find}
}
\description{
Get k nearest neighbors for CODEX protein space. If CleanCODEX
has been run, uses cleaned CODEX protein data, otherwise uses
original CODEX protein data.
}
