% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MapCODEXtoCITE.internal}
\alias{MapCODEXtoCITE.internal}
\title{Run all methods for anchor correction to map the CODEX dataset to the CITE-seq dataset
Since the CODEX dataset is usually much bigger, allows to subsample equal-sized
chunks of the CODEX dataset to map individually
Takes matrices and data frames instead of STvEA.data class}
\usage{
MapCODEXtoCITE.internal(cite_protein, codex_protein, cite_latent, num_chunks,
  seed = NULL, num_cores = 1, num.cc = NULL, k.anchor = 20,
  k.filter = 100, k.score = 80, k.weight = 100)
}
\arguments{
\item{cite_protein}{a (n cell x f feature) protein expression matrix}

\item{codex_protein}{a (m cell x f feature) protein expression matrix to be corrected}

\item{cite_latent}{a (cell x feature) embedding of the mRNA expression matrix from CITE-seq}

\item{num_chunks}{number of equal sized chunks to split CODEX dataset into for correction}

\item{seed}{set.seed before randomly sampling chunks of CODEX dataset}

\item{num_cores}{number of cores to use in parallelized correction of CODEX dataset.
On Windows, this must be set to 1.}

\item{num.cc}{number of canonical vectors to calculate. Defaults to number of proteins - 1}

\item{k.anchor}{number of nn used to find anchors via mutual nearest neighbors}

\item{k.filter}{number of nn in original feature space to use for filtering}

\item{k.score}{number of nn to use in shared nearest neighbor scoring}

\item{k.weight}{number of nn in original query feature space to make correction vectors}
}
\value{
a (m cell x f feature) expression matrix of the CODEX data corrected into the CITE-seq space
}
\description{
Run all methods for anchor correction to map the CODEX dataset to the CITE-seq dataset
Since the CODEX dataset is usually much bigger, allows to subsample equal-sized
chunks of the CODEX dataset to map individually
Takes matrices and data frames instead of STvEA.data class
}
