% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MapCODEXtoCITE}
\alias{MapCODEXtoCITE}
\title{Wraps MapCODEXtoCITE.internal using STvEA.data class}
\usage{
MapCODEXtoCITE(stvea_object, num_chunks, seed = NULL, num_cores = 1,
  num.cc = 29, k.anchor = 20, k.filter = 100, k.score = 80,
  k.weight = 100)
}
\arguments{
\item{stvea_object}{STvEA.data class with CODEX and CITE-seq data}

\item{num_chunks}{number of equal sized chunks to split CODEX dataset into for correction}

\item{seed}{set.seed before randomly sampling chunks of CODEX dataset}

\item{num_cores}{number of cores to use in parallelized correction of CODEX dataset}

\item{num.cc}{number of canonical vectors to calculate}

\item{k.anchor}{number of nn used to find anchors via mutual nearest neighbors}

\item{k.filter}{number of nn in original feature space to use for filtering}

\item{k.score}{number of nn to use in shared nearest neighbor scoring}

\item{k.weight}{number of nn in original query feature space to make correction vectors}
}
\description{
Wraps MapCODEXtoCITE.internal using STvEA.data class
}
