% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{MappingDataHolder-class}
\alias{MappingDataHolder-class}
\alias{MappingDataHolder}
\title{Class to hold all data matrices needed for mapping CODEX to CITE-seq}
\description{
Class to hold all data matrices needed for mapping CODEX to CITE-seq
}
