% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{NormalizeCITE.internal}
\alias{NormalizeCITE.internal}
\title{Normalize CITE-seq ADT expression by original cell expression totals
Takes matrices and data frames instead of STvEA.data class}
\usage{
NormalizeCITE.internal(cite_protein_clean, cite_protein_raw, batch = rep(1,
  nrow(cite_protein_raw)), k.batch = 20)
}
\arguments{
\item{cite_protein_clean}{Cleaned CITE-seq protein data (cell x protein) output by CleanCITE()}

\item{cite_protein_raw}{Raw CITE-seq protein data (cell x protein)}

\item{batch}{Batch indices for running MNN batch correction}

\item{k.batch}{k to use for MNN batch correction}
}
\value{
Normalized and batch corrected CITE-seq protein data matrix (cell x protein)
}
\description{
Normalize CITE-seq ADT expression by original cell expression totals
Takes matrices and data frames instead of STvEA.data class
}
