% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{NormalizeCITE}
\alias{NormalizeCITE}
\title{Wraps NormalizeCITE.internal using MappingDataHolder class}
\usage{
NormalizeCITE(mapping_object)
}
\arguments{
\item{mapping_object}{MappingDataHOlder class with CITE-seq protein data after CleanCITE}
}
\description{
Wraps NormalizeCITE.internal using MappingDataHolder class
}
