% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{NormCells}
\alias{NormCells}
\title{Normalize protein matrix by dividing out total counts per cell,
accounting for cells with no counts by keeping them 0
Usually to_norm and norm_by will be the same count matrix,
but sometimes we want to normalize processed data by raw}
\usage{
NormCells(to_norm, norm_by = to_norm)
}
\arguments{
\item{to_norm}{a (cells x proteins) matrix}

\item{norm_by}{a (cells x proteins) matrix}
}
\description{
Normalize protein matrix by dividing out total counts per cell,
accounting for cells with no counts by keeping them 0
Usually to_norm and norm_by will be the same count matrix,
but sometimes we want to normalize processed data by raw
}
