% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mrna_spatial}
\alias{plot_mrna_spatial}
\title{Plot CITE-seq gene expression on
CODEX spatial coordinates}
\usage{
plot_mrna_spatial(gene_name)
}
\description{
Plot CITE-seq gene expression on
CODEX spatial coordinates
}
