% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_neighbors_spatial}
\alias{plot_neighbors_spatial}
\title{Plot the CODEX neighbors of a set of CITE-seq cells
in the spatial coordinates}
\usage{
plot_neighbors_spatial(cite_cells)
}
\description{
Plot the CODEX neighbors of a set of CITE-seq cells
in the spatial coordinates
}
