% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_neighbors_umap}
\alias{plot_neighbors_umap}
\title{Plot CITE-seq neighbors of a set of CODEX cells
in the CITE-seq mRNA UMAP}
\usage{
plot_neighbors_umap(codex_cells)
}
\description{
Plot CITE-seq neighbors of a set of CODEX cells
in the CITE-seq mRNA UMAP
}
