% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotClusterCITE}
\alias{PlotClusterCITE}
\title{Plot clusters of CITE-seq cells in CITE-seq UMAP or t-SNE embedding}
\usage{
PlotClusterCITE(stvea_object, pt_size = 0.5)
}
\arguments{
\item{stvea_object}{STvEA.data class with CITE-seq embedding and clustering}
}
\description{
Plot clusters of CITE-seq cells in CITE-seq UMAP or t-SNE embedding
}
