% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotClusterCODEXemb}
\alias{PlotClusterCODEXemb}
\title{Plot clusters of CODEX cells in CODEX UMAP embedding}
\usage{
PlotClusterCODEXemb(stvea_object, pt_size = 0.5)
}
\arguments{
\item{stvea_object}{STvEA.data class with CODEX embedding and clustering}
}
\description{
Plot clusters of CODEX cells in CODEX UMAP embedding
}
