% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotExprCITE}
\alias{PlotExprCITE}
\title{Plot expression of a gene or protein in the CITE-seq UMAP embedding}
\usage{
PlotExprCITE(stvea_object, name, type = "RNA")
}
\arguments{
\item{stvea_object}{STvEA.data class with CITE-seq expression and embedding}

\item{name}{gene or protein name to plot}

\item{type}{which type of expression data should be plotted? "RNA" or "protein"}
}
\description{
Plot expression of a gene or protein in the CITE-seq UMAP embedding
}
