% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotExprCODEXemb}
\alias{PlotExprCODEXemb}
\title{Plot expression of a gene or protein in the CODEX UMAP or t-SNE embedding}
\usage{
PlotExprCODEXemb(stvea_object, name, type = "protein", high_color = "red",
  high_color2 = "green", low_color = "light gray", pt_size = 0.8)
}
\arguments{
\item{stvea_object}{STvEA.data class with CODEX expression and embedding}

\item{name}{gene or protein name to plot}

\item{type}{which type of expression data should be plotted? "RNA" or "protein"}

\item{high_color}{color of high expression on the color ramp}

\item{high_color2}{color of high expression for the second protein or gene, if provided}

\item{low_color}{color of low/no expression on the color ramp}

\item{pt_size}{ggplot geom_point size}
}
\description{
Plot expression of a gene or protein in the CODEX UMAP or t-SNE embedding
}
