% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotExprCODEXspatial}
\alias{PlotExprCODEXspatial}
\title{Plot expression of a gene or protein in the CITE-seq spatial coordinates}
\usage{
PlotExprCODEXspatial(stvea_object, name, type = "protein")
}
\arguments{
\item{stvea_object}{STvEA.data class with CITE-seq expression and spatial xy}

\item{name}{gene or protein name to plot}

\item{type}{which type of expression data should be plotted? "RNA" or "protein"}
}
\description{
Plot expression of a gene or protein in the CITE-seq spatial coordinates
}
