% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotExprCODEXumap}
\alias{PlotExprCODEXumap}
\title{Plot expression of a gene or protein in the CODEX UMAP embedding}
\usage{
PlotExprCODEXumap(stvea_object, name, type = "protein")
}
\arguments{
\item{stvea_object}{STvEA.data class with CODEX expression and embedding}

\item{name}{gene or protein name to plot}

\item{type}{which type of expression data should be plotted? "RNA" or "protein"}
}
\description{
Plot expression of a gene or protein in the CODEX UMAP embedding
}
