% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotIndexCODEXemb}
\alias{PlotIndexCODEXemb}
\title{Plot location of a CODEX cell in the CODEX UMAP or t-SNE embedding}
\usage{
PlotIndexCODEXemb(stvea_object, index, high_color = "red",
  low_color = "gray", pt_size = 0.8)
}
\arguments{
\item{stvea_object}{STvEA.data class with CODEX embedding}

\item{index}{index of CODEX cell to plot}

\item{high_color}{color of cell of interest}

\item{low_color}{color of all other cells}

\item{pt_size}{ggplot geom_point size}
}
\description{
Plot location of a CODEX cell in the CODEX UMAP or t-SNE embedding
}
