% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotIndexCODEXumap}
\alias{PlotIndexCODEXumap}
\title{Plot location of a CODEX cell in the CODEX UMAP embedding}
\usage{
PlotIndexCODEXumap(stvea_object, index)
}
\arguments{
\item{stvea_object}{STvEA.data class with CODEX embedding}

\item{index}{index of CODEX cell to plot}
}
\description{
Plot location of a CODEX cell in the CODEX UMAP embedding
}
