% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotSpatialNeighbors}
\alias{PlotSpatialNeighbors}
\title{Plot the CODEX neighbors of a set of CITE-seq cells
in the spatial coordinates}
\usage{
PlotSpatialNeighbors(cite_cells)
}
\description{
Plot the CODEX neighbors of a set of CITE-seq cells
in the spatial coordinates
}
