% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotSpatialRNA}
\alias{PlotSpatialRNA}
\title{Plot CITE-seq gene expression on
CODEX spatial coordinates}
\usage{
PlotSpatialRNA(gene_name)
}
\description{
Plot CITE-seq gene expression on
CODEX spatial coordinates
}
