% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotUmapNeighbors}
\alias{PlotUmapNeighbors}
\title{Plot CITE-seq neighbors of a set of CODEX cells
in the CITE-seq mRNA UMAP}
\usage{
PlotUmapNeighbors(codex_cells)
}
\description{
Plot CITE-seq neighbors of a set of CODEX cells
in the CITE-seq mRNA UMAP
}
