% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ReadDataFCS}
\alias{ReadDataFCS}
\title{Read in FCS file}
\usage{
ReadDataFCS(
  path_to_fcs,
  is_protein,
  is_blank,
  protein_names = NULL,
  tile_relative = FALSE,
  num_tile_x = 9,
  x_pix_size = 188,
  y_pix_size = 188,
  z_pix_size = 900,
  stvea_object = NULL
)
}
\arguments{
\item{path_to_fcs}{readable path to FCS file}

\item{is_protein}{boolean vector indicating which columns in FCS
should be kept as protein expression for downstream analyses}

\item{is_blank}{boolean vector indicating which columns in FCS
are blank for filtering purposes}

\item{protein_names}{vector of names for each protein channel column -
if null, will keep column names from FCS}

\item{tile_relative}{whether x,y,z coordinates are relative to x_tile,y_tile (TRUE)
or to corner of image (FALSE).
      If TRUE, will compute absolute coordinates
      If FALSE, will just take given x,y,z coordinates}

\item{num_tile_x}{number of tiles in the x direction, only used if tile_relative = TRUE
and the tile information in the FCS file is saved as tile_nr instead of x_tile,y_tile}

\item{x_pix_size}{x dimension of pixel coordinates in nm. Set to 1 to keep pixel dimensions.}

\item{y_pix_size}{y dimension of pixel coordinates in nm. Set to 1 to keep pixel dimensions.}

\item{z_pix_size}{z dimension of pixel coordinates in nm. Set to 1 to keep pixel dimensions.}

\item{stvea_object}{(optional) Pre-existing STvEA.data object to load data into.
If not provided, a new object is created.}
}
\value{
STvEA.data class object
}
\description{
Read in FCS file
}
