% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ReadNamesFCS}
\alias{ReadNamesFCS}
\title{Reads channel names from FCS two different ways}
\usage{
ReadNamesFCS(path_to_fcs)
}
\arguments{
\item{path_to_fcs}{readable path to FCS file}
}
\value{
list where "columns" are the column names from the FCS data matrix,
and "channels" are the readable names in the description
}
\description{
Reads channel names from FCS two different ways
}
