% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_anchor_correction.R
\name{RunCCA}
\alias{RunCCA}
\title{This file contains functions taken directly from the Seurat V3 source code,
edited to take matrices as input rather than Seurat objects and to be run on
only two datasets}
\usage{
RunCCA(
  object1,
  object2,
  standardize = FALSE,
  l2.norm = FALSE,
  num.cc = 20,
  setseed = TRUE,
  ...
)
}
\arguments{
\item{object1}{a (k features x n cells) matrix}

\item{object2}{a (k features x m cells) matrix}

\item{standardize}{Standardize matrices - scales columns to have unit variance
and mean 0}

\item{num.cc}{Number of canonical vectors to calculate}

\item{verbose}{...}
}
\value{
a ((n+m) cells x num.cc features) matrix
}
\description{
Modified from Seurat package
}
\references{
Stuart et al, bioRxiv 2018

\url{https://github.com/satijalab/seurat/tree/release/3.0}
Uses CCA to merge two datasets into a shared space
}
