% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{RunHDBSCAN}
\alias{RunHDBSCAN}
\title{Run HDBSCAN over parameter ranges
Must use Python HDBSCAN because it has 2 important parameters
(min_cluster_size and min_samples) but R HDBSCAN only has one (minPts)}
\usage{
RunHDBSCAN(cite_latent, umap_latent, min_cluster_size_range, min_sample_range)
}
\description{
Run HDBSCAN over parameter ranges
Must use Python HDBSCAN because it has 2 important parameters
(min_cluster_size and min_samples) but R HDBSCAN only has one (minPts)
}
