% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_anchor_correction.R
\name{ScoreAnchors}
\alias{ScoreAnchors}
\title{Score anchors between 0 and 1 based on shared nearest neighbors}
\usage{
ScoreAnchors(neighbors, anchors, num_cells_ref, num_cells_query, k.score = 30,
  verbose = FALSE)
}
\arguments{
\item{neighbors}{list of neighbors in CCA space as output by FindNN}

\item{anchors}{list of anchors as output by FindAnchorPairs or FilterAnchors}

\item{num_cells_ref}{total number of cells in dataset1}

\item{num_cells_query}{total number of cells in dataset2}

\item{k.score}{number of nn to use in shared nearest neighbor scoring}

\item{verbose}{...}
}
\value{
a matrix where each row is an anchor pair of cell indices with its score
}
\description{
Score anchors between 0 and 1 based on shared nearest neighbors
}
