% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SetData}
\alias{SetData}
\title{Check input data and create STvEA.data object}
\usage{
SetData(cite_protein, cite_latent, codex_protein, codex_size, codex_blanks)
}
\arguments{
\item{cite_protein}{Raw expression data for CITE-seq proteins (cell x protein)}

\item{cite_latent}{Low dimensional latent space for CITE-seq mRNA (cell x dim)}

\item{codex_protein}{Segmented and spillover corrected CODEX protein expression (cell x protein)}

\item{codex_size}{Size of each CODEX cell (vector)}

\item{codex_blanks}{Segmented and spillover corrected CODEX blank channel expression (cell x blank channel)}
}
\description{
Check input data and create STvEA.data object
}
