% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SetDataCITE}
\alias{SetDataCITE}
\title{Set CITE-seq data in STvEA.data object}
\usage{
SetDataCITE(
  cite_mRNA,
  cite_protein,
  cite_mRNA_norm = NULL,
  cite_latent = NULL,
  stvea_object = NULL
)
}
\arguments{
\item{cite_mRNA}{Raw count data for CITE-seq mRNA (cell x gene)}

\item{cite_protein}{Raw expression data for CITE-seq proteins (cell x protein)}

\item{cite_mRNA_norm}{(optional) Normalized and scaled CITE-seq mRNA count data. (cell x gene)
If not provided, raw count data is divided by total per cell, then log transformed.}

\item{cite_latent}{(optional) Low dimensional latent space for CITE-seq mRNA (cell x dim)}

\item{stvea_object}{(optional) Pre-existing STvEA.data object to load data into.
If not provided, a new object is created.}
}
\description{
Set CITE-seq data in STvEA.data object
}
