% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SetDataCODEX}
\alias{SetDataCODEX}
\title{Set CODEX data in STvEA.data object}
\usage{
SetDataCODEX(codex_protein, codex_blanks, codex_size, codex_spatial,
  stvea_object = NULL)
}
\arguments{
\item{codex_protein}{Segmented and spillover corrected CODEX protein expression (cell x protein)}

\item{codex_blanks}{Segmented and spillover corrected CODEX blank channel expression (cell x blank channel)}

\item{codex_size}{Size of each CODEX cell (vector)}

\item{codex_spatial}{xyz coordinates of each CODEX cell (cell x 3 coordinates)}

\item{stvea_object}{(optional) Pre-existing STvEA.data object to load data into.
If not provided, a new object is created.}
}
\description{
Set CODEX data in STvEA.data object
}
