% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{SSE}
\alias{SSE}
\title{Calculates the sum of squared errors in binned probabilities of count data}
\usage{
SSE(args, p_obs)
}
\arguments{
\item{args}{arguments used in the negative binomial mixture model}

\item{p_obs}{a named vector of the probabilities of observing a given count
in gene expression data, as output by running table() on the gene count data}
}
\description{
Calculates the sum of squared errors in binned probabilities of count data
}
