% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STvEA-package.R
\docType{package}
\name{STvEA-package}
\alias{STvEA}
\alias{STvEA-package}
\title{STvEA: Spatially-resolved Transcriptomics via Epitope Anchoring}
\description{
STvEA is an analysis pipeline for cleaning, clustering, and plotting CODEX and CITE-seq protein data, mapping a CODEX dataset to a matching CITE-seq dataset, and assessing colocalization of features using the Adjacency Score.
}
\author{
\strong{Maintainer}: Kiya W. Govek \email{govekk@pennmedicine.upenn.edu}

}
\keyword{internal}
