% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{STvEA.data-class}
\alias{STvEA.data-class}
\alias{STvEA.data}
\title{Class to hold all data matrices needed for mapping CODEX to CITE-seq
and following analysis}
\description{
Class to hold all data matrices needed for mapping CODEX to CITE-seq
and following analysis
}
