% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{TransferDataSeurat2}
\alias{TransferDataSeurat2}
\title{Transfer data from a Seurat object to STvEA.data object}
\usage{
TransferDataSeurat2(seurat_object, embedding_reduction = NULL,
  latent_reduction = NULL, latent_dims = 20)
}
\arguments{
\item{seurat_object}{Seurat class object}

\item{latent_dims}{}
}
\value{
STvEA.data class object
}
\description{
Transfer data from a Seurat object to STvEA.data object
}
