% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_anchor_correction.R
\name{TransformDataMatrix}
\alias{TransformDataMatrix}
\title{Correct the query dataset by subtracting the weighted anchor vectors}
\usage{
TransformDataMatrix(
  ref_mat,
  query_mat,
  integration.matrix,
  weights,
  verbose = FALSE
)
}
\arguments{
\item{ref_mat}{a (cell x feature) protein expression matrix}

\item{query_mat}{a (cell x feature) protein expression matrix to be corrected}

\item{integration.matrix}{matrix of anchor vectors (output of FindIntegrationMatrix)}

\item{weights}{weights of the anchors for each query cell (output of FindWeights)}

\item{verbose}{...}
}
\value{
a (cell x feature) corrected data matrix
   the original reference data appended to the corrected query data
}
\description{
Correct the query dataset by subtracting the weighted anchor vectors
}
