% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scores.R
\name{AdjScoreProteins.internal}
\alias{AdjScoreProteins.internal}
\title{Use the Adjacency Score to evaluate colocalization of given
pairs of proteins in the CODEX spatial dimensions
Takes matrices and data frames instead of STvEA.data class}
\usage{
AdjScoreProteins.internal(adj_matrix, codex_protein, protein_pairs, c = 0,
  num_cores = 1, num_perms = 1000, perm_estimate = T)
}
\arguments{
\item{adj_matrix}{a (preferrably sparse) binary matrix of
adjacency between the cells in the CODEX spatial coordinates}

\item{codex_protein}{a (cells x proteins) matrix of CODEX protein expression}

\item{protein_pairs}{a 2 column matrix of protein pairs where each row
specifies the names of the proteins in a pair.}

\item{c}{constant used to determine width of diffusion, must be 0 <= c}

\item{num_cores}{integer specifying the number of cores to be used
in the computation. By default only one core is used.
On Windows, this must be set to 1.}

\item{num_perms}{number of permutations used to build the null
distribution for each feature. By default is set to 1000.}

\item{perm_estimate}{boolean indicating whether Gaussian distribution
parameters should be determined from num_perms permutations to estimate
the p-value. By default is set to TRUE.}
}
\description{
Use the Adjacency Score to evaluate colocalization of given
pairs of proteins in the CODEX spatial dimensions
Takes matrices and data frames instead of STvEA.data class
}
