% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{CleanCITE.gaussian.internal}
\alias{CleanCITE.gaussian.internal}
\title{Removes noise from CITE-seq protein data by fitting
a two-component Gaussian mixture to the log-normalized
expression with the zeros removed, and then computing each
expression measurement as the cumulative distribution
of the Gaussian with the higher median.}
\usage{
CleanCITE.gaussian.internal(norm_cite_protein, num_cores = 1)
}
\arguments{
\item{norm_cite_protein}{Log-normalized CITE-seq protein data (cell x protein)}

\item{num_cores}{number of cores to use for parallelized fits.
On Windows, this must be set to 1.}
}
\value{
Cleaned CITE-seq protein data matrix (cell x protein)
}
\description{
This model is similar to the one proposed by Trong et al. in the SISUA preprint
(https://www.biorxiv.org/content/10.1101/631382v1)
}
\details{
Takes matrices and data frames instead of STvEA.data class
}
