% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{CleanCITE}
\alias{CleanCITE}
\title{Wraps CleanCITE.internal using MappingDataHolder class}
\usage{
CleanCITE(mapping_object, maxit = 500, factr = 1e-09,
  optim_inits = NULL, num_cores = 1)
}
\arguments{
\item{mapping_object}{MappingDataHolder class with CITE-seq protein data}

\item{maxit}{maximum number of iterations for optim function}

\item{factr}{accuracty of optim function}

\item{optim_inits}{a matrix of (proteins x params) with initialization
parameters for each protein to input to the optim function. If NULL,
starts at two default parameter sets and picks the better one}

\item{num_cores}{number of cores to use for parallelized fits}
}
\description{
Wraps CleanCITE.internal using MappingDataHolder class
}
