% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{CleanCODEX}
\alias{CleanCODEX}
\title{Wraps CleanCODEX.internal using STvEA.data class}
\usage{
CleanCODEX(stvea_object)
}
\arguments{
\item{stvea_object}{STvEA.data class with CODEX data after FilterCODEX}
}
\description{
Wraps CleanCODEX.internal using STvEA.data class
}
