% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{CleanCODEX}
\alias{CleanCODEX}
\title{Wraps CleanCODEX.internal using MappingDataHolder class}
\usage{
CleanCODEX(mapping_object)
}
\arguments{
\item{mapping_object}{MappingDataHolder class with CODEX data after FilterCODEX}
}
\description{
Wraps CleanCODEX.internal using MappingDataHolder class
}
