% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ClusterCODEX}
\alias{ClusterCODEX}
\title{Calls ClusterCODEX.internal with STvEA.data object}
\usage{
ClusterCODEX(stvea_object, k = NULL)
}
\arguments{
\item{stvea_object}{STvEA.data class object with KNN matrix for CODEX protein space}

\item{k}{number of nearest neighbors to use in graph-based clustering
If NULL, will use all nearest neighbors from the KNN matrix.}
}
\description{
Calls ClusterCODEX.internal with STvEA.data object
}
