% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{CorrectChunksCODEX}
\alias{CorrectChunksCODEX}
\title{Wraps CorrectChunksCODEX.internal using MappingDataHolder class}
\usage{
CorrectChunksCODEX(mapping_object, num_chunks, seed = NULL,
  num.cc = 29, k.anchor = 20, k.filter = 100, k.score = 80,
  k.weight = 100)
}
\arguments{
\item{mapping_object}{MappingDataHolder class with CODEX and CITE-seq data}

\item{num_chunks}{number of equal sized chunks to split CODEX dataset into for correction}

\item{num.cc}{number of canonical vectors to calculate}

\item{k.anchor}{number of nn used to find anchors via mutual nearest neighbors}

\item{k.filter}{number of nn in original feature space to use for filtering}

\item{k.score}{number of nn to use in shared nearest neighbor scoring}

\item{k.weight}{number of nn in original query feature space to make correction vectors}
}
\description{
Wraps CorrectChunksCODEX.internal using MappingDataHolder class
}
