% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{NormalizeCITE}
\alias{NormalizeCITE}
\title{Normalize CITE-seq ADT expression by original cell expression totals}
\usage{
NormalizeCITE(stvea_object)
}
\arguments{
\item{stvea_object}{STvEA.data class with CITE-seq protein data after CleanCITE}
}
\description{
Normalize CITE-seq ADT expression by original cell expression totals
}
