% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{NormalizeCITE}
\alias{NormalizeCITE}
\title{Wraps NormalizeCITE.internal using STvEA.data class}
\usage{
NormalizeCITE(stvea_object)
}
\arguments{
\item{stvea_object}{STvEA.data class with CITE-seq protein data after CleanCITE}
}
\description{
Wraps NormalizeCITE.internal using STvEA.data class
}
