% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotIndexCITE}
\alias{PlotIndexCITE}
\title{Plot location of a CITE-seq cell in the CITE-seq UMAP embedding}
\usage{
PlotIndexCITE(stvea_object, index)
}
\arguments{
\item{stvea_object}{STvEA.data class with CITE-seq embedding}

\item{index}{index of CITE-seq cell to plot}
}
\description{
Plot location of a CITE-seq cell in the CITE-seq UMAP embedding
}
