% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{TransferDataSeurat2}
\alias{TransferDataSeurat2}
\title{Transfer data from a Seurat object to STvEA.data object}
\usage{
TransferDataSeurat2(
  seurat_object,
  embedding_reduction = NULL,
  latent_reduction = NULL,
  latent_dims = 20
)
}
\arguments{
\item{seurat_object}{Seurat class object}

\item{embedding_reduction}{(optional) name of the slot holding the dimension reduction
embedding used for visualization. Examples: "umap", "tsne"}

\item{latent_reduction}{(optional) name of the lot holding the dimension reduction to
be used for the CITE-seq mRNA latent space. Examples: "pca"}

\item{latent_dims}{number of dimensions to use for the latent space,
not used if latent_reduction is not defined}
}
\value{
STvEA.data class object
}
\description{
Transfer data from a Seurat object to STvEA.data object
}
